Move settings to a preferencefragmentcompat Add syncing amiibo info from amiiboapi.com Add image url to Amiibo Use Snackbar in settings
